# openapi/enterprise-video-factory.openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Video Factory API
  version: 0.1.0
  description: Draft resource model for enterprise automated talking-head video generation.
servers:
  - url: https://api.example.local
security:
  - bearerAuth: []
paths:
  /v1/auth/login:
    post:
      summary: Authenticate a Studio user
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
...
  /v1/copywriter/generate:
    post:
      summary: Generate social-ready talking-head copy
      operationId: generateCopywriterScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopywriterGenerateRequest'
      responses:
        '200':
          description: Generated copywriting result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CopywriterResult'
  /v1/jobs:
    get:
      summary: List video generation jobs
      operationId: listJobs
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
    post:
      summary: Create a video generation job
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
  /v1/jobs/{job_id}:
    get:
      summary: Get job state
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: Job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
  /v1/jobs/{job_id}/start:
    post:
      summary: Start a video generation job
      operationId: startJob
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: Job state after generation start
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...
  /v1/jobs/{job_id}/artifacts:
    get:
      summary: List job artifacts
      operationId: listJobArtifacts
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: Artifact list
          content:
            application/json:
              schema:
                type: object
                required: [artifacts]
                properties:
                  artifacts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Artifact'
  /v1/jobs/{job_id}/qc:
    get:
      summary: List job QC reports
      operationId: listJobQcReports
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: QC reports
          content:
            application/json:
              schema:
                type: object
                required: [qc_reports]
                properties:
                  qc_reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/QcReport'
...
  /v1/assets:
    get:
      summary: List assets
      operationId: listAssets
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                type: object
                required: [assets]
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
    post:
      summary: Create an asset draft
      operationId: createAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
...
  /v1/assets/{asset_id}/signed-url:
    get:
      summary: Create a signed tenant-scoped asset media URL
      operationId: getAssetSignedUrl
      parameters:
        - $ref: '#/components/parameters/assetId'
      responses:
        '200':
          description: Signed asset media URL
          content:
            application/json:
              schema:
                type: object
                required: [url, expires_at]
                properties:
                  url:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
  /v1/assets/{asset_id}/media:
    get:
      summary: Serve signed local asset media
      operationId: getAssetMedia
      parameters:
        - $ref: '#/components/parameters/assetId'
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asset media bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/mp4:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
            video/quicktime:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
  /v1/assets/{asset_id}/consent:
    post:
      summary: Attach consent evidence to an asset
      operationId: attachConsent
      parameters:
        - $ref: '#/components/parameters/assetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
      responses:
        '201':
          description: Consent event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentEvent'
  /v1/assets/{asset_id}/activate:
    post:
      summary: Activate an authorized asset
      operationId: activateAsset
      parameters:
        - $ref: '#/components/parameters/assetId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetStateChange'
      responses:
        '200':
          description: Asset activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
...
  /v1/artifacts/{artifact_id}/signed-url:
    get:
      summary: Create a signed tenant-scoped artifact URL
      operationId: getArtifactSignedUrl
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Signed URL
          content:
            application/json:
              schema:
                type: object
                required: [url, expires_at]
                properties:
                  url:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    jobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
    assetId:
      name: asset_id
      in: path
      required: true
      schema:
        type: string
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthSession:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
        token_type:
          enum: [bearer]
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [user_id, tenant_id, email, role]
      properties:
        user_id:
          type: string
        tenant_id:
          type: string
        email:
          type: string
          format: email
        role:
          enum: [creator, tenant_admin, reviewer, auditor]
...
    CopywriterGenerateRequest:
      type: object
      required: [situation]
      properties:
        situation:
          type: string
        audience:
          type: string
        pain:
          type: string
        offer:
          type: string
        proof:
          type: string
        action:
          type: string
        tone:
          type: string
        duration_sec:
          type: integer
          enum: [30, 45, 60]
    CopywriterResult:
      type: object
      required: [mode, title, hook, script, beats, caption, hashtags, notes]
      properties:
        mode:
          enum: [openai, local_template, openai_failed_local_fallback]
        title:
          type: string
        hook:
          type: string
        script:
          type: string
        beats:
          type: array
          items:
            type: string
        caption:
          type: string
        hashtags:
          type: array
          items:
            type: string
        notes:
          type: array
          items:
            type: string
        provider_error:
          type: string
    JobState:
      type: string
      enum: [created, validated, planning, planned, segmenting, generating_audio, generating_avatar, rendering_scenes, aligning_subtitles, mixing_audio, composing, quality_checking, repairing, review_required, publish_ready, failed, cancelled]
...
    ManifestStage:
      type: string
      enum: [source_intake, planning, segment_input, tts, avatar, alignment, scene_render, audio_mix, composition, qc, repair, review, release]
    QcStatus:
      type: string
      enum: [passed, warning, failed_repairable, review_required, failed_terminal]
    FailureCode:
      type: string
      enum: [asset_not_authorized, consent_expired, tenant_scope_violation, scene_dsl_invalid, scene_dsl_prohibited_code, tts_text_too_long, voice_profile_missing, model_runtime_error, renderer_runtime_error, audio_unreadable, language_unsupported, audio_clipped, audio_silence, asr_text_mismatch, alignment_failed, avatar_profile_missing, face_not_detected, lip_sync_below_threshold, avatar_identity_drift, template_not_approved, render_blank, text_overflow, subtitle_safe_area_failed, missing_segment, duration_mismatch, codec_failed, audio_mix_failed, output_corrupt, traceability_incomplete, manual_review_required, model_license_unverified]
...
    AssetState:
      type: string
      enum: [draft, pending_consent, consent_verified, active, suspended, revoked, archived]
...
    CreateJobRequest:
      type: object
      required: [input, asset_refs, output_profile]
      properties:
        input:
          type: object
          required: [type, text]
          properties:
            type:
              enum: [text, article, opinion]
            text:
              type: string
        asset_refs:
          type: object
          required: [voice_profile_id, avatar_profile_id]
          properties:
            voice_profile_id:
              type: string
            avatar_profile_id:
              type: string
            brand_kit_id:
              type: string
            music_asset_id:
              type: string
            template_id:
              type: string
        output_profile:
          type: object
          required: [duration_target_sec, aspect_ratio]
          properties:
            duration_target_sec:
              type: integer
              minimum: 30
              maximum: 90
            aspect_ratio:
              enum: ['9:16']
    Job:
      type: object
      required: [job_id, tenant_id, state]
      properties:
        job_id:
          type: string
        tenant_id:
          type: string
        state:
          $ref: '#/components/schemas/JobState'
        active_stage:
          $ref: '#/components/schemas/ManifestStage'
        asset_authorization_snapshots:
          type: array
          items:
            type: string
        asset_bindings:
          type: array
          items:
            $ref: '#/components/schemas/AssetBinding'
        progress:
          $ref: '#/components/schemas/JobProgress'
        state_history:
          type: array
          items:
            type: string
    JobProgress:
      type: object
      required: [segments_total, segments_completed, repair_attempts]
      properties:
        segments_total:
          type: integer
        segments_completed:
          type: integer
        repair_attempts:
          type: integer
    AssetBinding:
      type: object
      required: [request_field, asset_id, asset_type, role]
      properties:
        request_field:
          type: string
        asset_id:
          type: string
        asset_type:
          enum: [voice_profile, avatar_profile, brand_kit, music_asset, template]
        role:
          type: string
        authorization_snapshot:
          type: string
...
    Artifact:
      type: object
      required: [artifact_id, job_id, stage, manifest_uri]
      properties:
        artifact_id:
          type: string
        job_id:
          type: string
        segment_id:
          type: string
        stage:
          $ref: '#/components/schemas/ManifestStage'
        attempt:
          type: integer
        manifest_uri:
          type: string
        qc_status:
          $ref: '#/components/schemas/QcStatus'
        signed_url_available:
          type: boolean
        file_uri:
          type: string
    QcReport:
      type: object
      required: [qc_report_id, job_id, stage, status]
      properties:
        qc_report_id:
          type: string
        job_id:
          type: string
        segment_id:
          type: string
        stage:
          $ref: '#/components/schemas/ManifestStage'
        status:
          $ref: '#/components/schemas/QcStatus'
        failure_codes:
          type: array
          items:
            $ref: '#/components/schemas/FailureCode'
        message:
          type: string
...
    Asset:
      type: object
      required: [asset_id, asset_type, state]
      properties:
        asset_id:
          type: string
        tenant_id:
          type: string
        asset_type:
          enum: [voice_profile, avatar_profile, brand_kit, music_asset, template]
        state:
          $ref: '#/components/schemas/AssetState'
        display_name:
          type: string
        authorization_snapshots:
          type: array
          items:
            type: string
        file_uri:
          type: string
    AssetStateChange:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    CreateAssetRequest:
      type: object
      required: [asset_type]
      properties:
        asset_type:
          enum: [voice_profile, avatar_profile, brand_kit, music_asset, template]
        display_name:
          type: string
        file_name:
          type: string
        content_base64:
          type: string
    ConsentRequest:
      type: object
      required: [scope, evidence_uri]
      properties:
        scope:
          type: object
        evidence_uri:
          type: string
    ConsentEvent:
      type: object
      required: [consent_event_id, asset_id, granted_at]
      properties:
        consent_event_id:
          type: string
        asset_id:
          type: string
        granted_at:
          type: string
          format: date-time
...